Map reduce example implemented

+++ model/src/main/java/experiments/interfaces/nikita/Grouping.java
@@ -5,6 +5,6 @@ import java.util.function.Function;
 /**
  * Created by marnikitta on 26.10.16.
  */
-public interface Grouping<S> extends Function<S, Long> {
+public interface Grouping<S> extends Function<S, Integer> {
     boolean equals(S s1, S s2);
 }
